+++ docs/changelog.rst
@@ -23,6 +23,7 @@ Changelog
 - Internal error when an API operation contains multiple parameters with the same name and some of them contain the ``examples`` keyword.
 - Internal error during query parameter generation on schemas that do not contain the ``type`` keyword.
 - Example generation for request body parameters using ``$ref``.
+- Generating examples for properties that have deeply nested ``$ref``. 
 
 .. _v3.24.3: